requirements.txt
# VRC Uploader Dependencies
PyQt6>=6.4.0
Pillow>=9.0.0
httpx>=0.23.0
watchdog>=2.1.0
qasync>=0.23.0
